fix: replace zendx date picker with bootstrap date picker
The forms edited have jQuery code to instantiate the date elements as a
bootstrap date picker. However, these date fields where set to be ZendX
_JQuery_Form_Element_DatePicker rather than a standard text field as
required.

Due to this, options set with the updated bootstrap date picker failed
to be set, while to ZendX date picker options failed to work.
