+++ app/modules/admin/views/scripts/quotes/add.phtml
@@ -5,7 +5,7 @@ $this->metaBase()->setDescription($this->title())->generate();
 ?>
 <?php
 $this->jQuery()->addJavascriptFile(
-    $this->baseUrl() . '/js/bootstrap-datepicker-v1.9.0.js',
+    $this->baseUrl() . '/js/bootstrap-datepicker.js',
     $type = 'text/javascript'
 );
 $this->headLink()->appendStylesheet('/css/bootstrap-datepicker.css');
