+++ app/modules/admin/views/scripts/research/add.phtml
@@ -8,7 +8,7 @@ $this->metaBase()->setDescription($this->title())->generate();
 $this->jQuery()->addJavascriptFile($this->baseUrl() . '/js/JQuery/jquery.autocomplete.pack.js', $type = 'text/javascript');
 $this->jQuery()->addJavascriptFile($this->baseUrl() . '/js/JQuery/autocompleteinitPeople.js', $type = 'text/javascript');
 $this->jQuery()->addJavascriptFile(
-    $this->baseUrl() . '/js/bootstrap-datepicker-v1.9.0.js',
+    $this->baseUrl() . '/js/bootstrap-datepicker.js',
     $type = 'text/javascript'
 );
 $this->headLink()->appendStylesheet('/css/bootstrap-datepicker.css');
