+++ app/modules/admin/views/scripts/research/edit.phtml
@@ -8,18 +8,27 @@ $this->metaBase()->setDescription($this->title())->generate();
 <?php
 $this->jQuery()->addJavascriptFile($this->baseUrl() . '/js/JQuery/jquery.autocomplete.pack.js', $type = 'text/javascript');
 $this->jQuery()->addJavascriptFile($this->baseUrl() . '/js/JQuery/autocompleteinitPeople.js', $type = 'text/javascript');
-$this->jQuery()->addJavascriptFile($this->baseUrl() . '/js/bootstrap-datepicker.js', $type = 'text/javascript');
+$this->jQuery()->addJavascriptFile(
+    $this->baseUrl() . '/js/bootstrap-datepicker.js',
+    $type = 'text/javascript'
+);
+$this->headLink()->appendStylesheet('/css/bootstrap-datepicker.css');
 $this->jQuery()->addJavascriptFile($this->baseUrl() . '/js/JQuery/jquery.are-you-sure.js', $type = 'text/javascript');
 $this->headLink()->appendStylesheet($this->baseUrl() . '/css/autocomplete.css');
 ?>
 <?php $this->jQuery()->javascriptCaptureStart() ?>
 $(document).ready(function() {
-$("#startDate").datepicker({ maxDate: '+6y'});
-$("#endDate").datepicker({ maxDate: '+6y' });
-$('form').areYouSure();
-$("form").submit(function() {
-$(":submit", this).attr("disabled", "disabled");
-});
+    var datePickerSettings = {
+        endDate: (new Date((new Date()).getFullYear() + 6, 12, 0)),
+        format: 'yyyy-mm-dd',
+    };
+    $("#startDate").datepicker(datePickerSettings);
+    $("#endDate").datepicker(datePickerSettings);
+
+    $('form').areYouSure();
+    $("form").submit(function() {
+        $(":submit", this).attr("disabled", "disabled");
+    });
 });
 
 <?php $this->jQuery()->javascriptCaptureEnd() ?>
