+++ app/modules/admin/views/scripts/users/upgrade.phtml
@@ -3,17 +3,29 @@ $this->headTitle('Approve an account');
 $this->metaBase()->setDescription('Approve an account for higher level access')->generate();
 ?>
 
-    <h2 class="lead"><?php echo $this->title(); ?></h2>
+<h2 class="lead"><?php echo $this->title(); ?></h2>
 <?php echo $this->form; ?>
 <?php
-$this->jQuery()->addJavascriptFile($this->baseUrl() . '/js/bootstrap-datepicker.js', $type = 'text/javascript');
+$this->jQuery()->addJavascriptFile(
+    $this->baseUrl() . '/js/bootstrap-datepicker.js',
+    $type = 'text/javascript'
+);
+$this->headLink()->appendStylesheet('/css/bootstrap-datepicker.css');
 $this->headLink()->appendStylesheet($this->baseUrl() . '/css/autocomplete.css');
 ?>
+
 <?php $this->jQuery()->javascriptCaptureStart(); ?>
 $(document).ready(function() {
-$('form').areYouSure();
-$("form").submit(function() {
-$(":submit", this).attr("disabled", "disabled");
-});
+    var datePickerSettings = {
+        format: 'yyyy-mm-dd'
+    };
+
+    $("#startDate").datepicker(datePickerSettings);
+    $("#endDate").datepicker(datePickerSettings);
+
+    $('form').areYouSure();
+    $("form").submit(function() {
+        $(":submit", this).attr("disabled", "disabled");
+    });
 });
 <?php $this->jQuery()->javascriptCaptureEnd(); ?>
