+++ app/modules/database/views/scripts/artefacts/add.phtml
@@ -15,7 +15,7 @@ $this->addCheck()->canThey();
 <?php echo $this->form; ?>
 <?php
 $this->jQuery()->addJavascriptFile(
-    $this->baseUrl() . '/js/bootstrap-datepicker-v1.9.0.js',
+    $this->baseUrl() . '/js/bootstrap-datepicker.js',
     $type = 'text/javascript'
 );
 $this->headLink()->appendStylesheet('/css/bootstrap-datepicker.css');
