+++ app/modules/database/views/scripts/hoards/add.phtml
@@ -31,7 +31,7 @@ echo $this->form;
 ?>
 <?php
 $this->jQuery()->addJavascriptFile(
-    $this->baseUrl() . '/js/bootstrap-datepicker-v1.9.0.js',
+    $this->baseUrl() . '/js/bootstrap-datepicker.js',
     $type = 'text/javascript'
 );
 $this->headLink()->appendStylesheet('/css/bootstrap-datepicker.css');
