+++ app/modules/database/views/scripts/rallies/add.phtml
@@ -3,8 +3,13 @@ $this->headTitle('Add a rally');
 $this->jQuery()->javascriptCaptureStart() ?>
 $(document).ready(function() {
 
-$("#date_from").datepicker({ maxDate: '+1y'});
-$("#date_to").datepicker({ maxDate: '+1y'});
+    var datePickerSettings = {
+        format: 'yyyy-mm-dd',
+        endDate: (new Date((new Date()).getFullYear() + 1, 12, 0)),
+    };
+
+    $("#date_from").datepicker(datePickerSettings);
+    $("#date_to").datepicker(datePickerSettings);
 
 });
 <?php $this->jQuery()->javascriptCaptureEnd();
@@ -23,6 +28,8 @@ $this->jQuery()->addJavascriptFile($this->baseUrl()
 
 $this->jQuery()->addJavascriptFile($this->baseUrl()
         .'/js/JQuery/typeaheads/organiser.js', $type='text/javascript');
+
+$this->headLink()->appendStylesheet('/css/bootstrap-datepicker.css');
 ?>
 
 <h2 class="lead"><?php echo $this->title();?></h2>
