+++ app/modules/database/views/scripts/rallies/addflo.phtml
@@ -1,5 +1,17 @@
 <?php 
-$this->jQuery()->addJavascriptFile($this->baseUrl() . '/js/bootstrap-datepicker.js',$type='text/javascript'); 
+$this->jQuery()->addJavascriptFile($this->baseUrl() . '/js/bootstrap-datepicker.js',$type='text/javascript');
+$this->headLink()->appendStylesheet('/css/bootstrap-datepicker.css');
 $this->headTitle('Add an attending Finds officer to a rally');?>
 <h2 class="lead"><?php echo $this->title();?></h2>
+<?php $this->jQuery()->javascriptCaptureStart() ?>
+    $(document).ready(function() {
+        var datePickerSettings = {
+        format: 'yyyy-mm-dd',
+        };
+
+        $("#dateFrom").datepicker(datePickerSettings);
+        $("#dateTo").datepicker(datePickerSettings);
+    });
+<?php $this->jQuery()->javascriptCaptureEnd(); ?>
+
 <?php echo $this->form;?>
