+++ app/modules/database/views/scripts/search/advanced.phtml
@@ -15,7 +15,7 @@
             . '/js/JQuery/typeaheads/secondaryID.js', $type='text/javascript');
 
     $this->jQuery()->addJavascriptFile($this->baseUrl()
-            . '/js/bootstrap-datepicker-v1.9.0.js',$type='text/javascript');
+            . '/js/bootstrap-datepicker.js',$type='text/javascript');
     $this->jQuery()->addJavascriptFile($this->baseUrl()
             . '/js/JQuery/jQueryLinkedSelect.js',$type='text/javascript');
     $this->jQuery()->addJavascriptFile($this->baseUrl()
