+++ app/modules/database/views/scripts/search/hoards.phtml
@@ -15,7 +15,7 @@
             . '/js/JQuery/typeaheads/secondaryID.js', $type='text/javascript');
 
     $this->jQuery()->addJavascriptFile(
-        $this->baseUrl() . '/js/bootstrap-datepicker-v1.9.0.js',
+        $this->baseUrl() . '/js/bootstrap-datepicker.js',
         $type = 'text/javascript'
     );
     $this->jQuery()->addJavascriptFile($this->baseUrl()
