+++ app/modules/database/views/scripts/statistics/county.phtml
@@ -282,7 +282,7 @@ $this->metaBase()->setDescription($this->title())->generate();
 <?php endif; ?>
 <?php
 $this->jQuery()->addJavascriptFile(
-    $this->baseUrl() . '/js/bootstrap-datepicker-v1.9.0.js',
+    $this->baseUrl() . '/js/bootstrap-datepicker.js',
     $type = 'text/javascript'
 );
 $this->headLink()->appendStylesheet('/css/bootstrap-datepicker.css');
