+++ app/modules/users/views/scripts/calendar/add.phtml
@@ -1,8 +1,18 @@
  <?php
 $this->jQuery()->addJavascriptFile($this->baseUrl() . '/js/bootstrap-datepicker.js',$type='text/javascript');
+$this->headLink()->appendStylesheet('/css/bootstrap-datepicker.css');
 $this->headTitle('Add a new shared calendar event');
 ?>
 <h2 class="lead">
     <?php echo $this->title();?>
 </h2>
+<?php $this->jQuery()->javascriptCaptureStart(); ?>
+    $(document).ready(function() {
+        var datePickerSettings = {
+            format: 'yyyy-mm-dd',
+        };
+        $("#startDate").datepicker(datePickerSettings);
+        $("#endDate").datepicker(datePickerSettings);
+    });
+<?php $this->jQuery()->javascriptCaptureEnd(); ?>
 <?php echo $this->form;?>
