+++ app/modules/users/views/scripts/vacancies/add.phtml
@@ -1,7 +1,7 @@
 <?php
 $this->headTitle('Add a new vacancy');
 $this->jQuery()->addJavascriptFile(
-    $this->baseUrl() . '/js/bootstrap-datepicker-v1.9.0.js',
+    $this->baseUrl() . '/js/bootstrap-datepicker.js',
     $type = 'text/javascript'
 );
 $this->headLink()->appendStylesheet('/css/bootstrap-datepicker.css');
